fix(ci): python-coverage-shards - hosted validation delegate tests exceeded line limit
